Fix issue with commit 8d2dc760b
The simple_rel_array cannot be set NULL and has to be updated by copying
forward the elements after compressed_rel->relid and setting the last
element to NULL. Few other updates needs to be done to that struct as
well. For now just don't set it to NULL.

Needs further investigation.